+++ lib/pliny/middleware/versioning.rb
@@ -27,7 +27,7 @@ module Pliny::Middleware
 Please specify a version along with the MIME type. For example, `Accept: application/vnd.#{@app_name}+json; version=1`.
             eos
             return [400, { "content-type" => "application/json; charset=utf-8" },
-              [MultiJson.encode(error)]]
+              [JSON.generate(error)]]
           end
 
           unless version