+++ lib/template/spec/endpoints/health_spec.rb
@@ -13,7 +13,7 @@ RSpec.describe Endpoints::Health do
       assert_equal(200, last_response.status)
       assert_equal("application/json;charset=utf-8", last_response.headers["Content-Type"])
       assert_equal(2, last_response.headers["Content-Length"].to_i)
-      assert_equal({}, MultiJson.decode(last_response.body))
+      assert_equal({}, JSON.parse(last_response.body))
     end
   end
 
@@ -49,7 +49,7 @@ RSpec.describe Endpoints::Health do
       assert_equal(200, last_response.status)
       assert_equal("application/json;charset=utf-8", last_response.headers["Content-Type"])
       assert_equal(2, last_response.headers["Content-Length"].to_i)
-      assert_equal({}, MultiJson.decode(last_response.body))
+      assert_equal({}, JSON.parse(last_response.body))
     end
   end
 end